requirements.txt
# Django
Django==5.2
djangorestframework==3.16.1
pillow==11.3.0
djangorestframework_simplejwt==5.5.1

# Authentication

# Database